+++ CHANGELOG.md
@@ -12,6 +12,15 @@ small and tractable.
 - Fixed bug in template form of initial state when mixing from above in non-template form.
 - Added alpha versions of new, tighter, DOM functions that do not require props, or #js
 - A few minor bug fixes in i18n alpha
+- PORTING TO fulcro.client.alpha.dom
+  - You cannot use expressions (e.g. list forms) for props. Only symbols, maps, or js objects
+  - Proper runtime optimization cannot be done for isomorphic rendering with a single namespace (this was a bug inherited from Om Next). For CLJC, the correct require is now:
+  ```
+  #?(:clj [fulcro.client.alpha.dom-server :as dom]
+     :cljs [fulcro.client.alpha.dom :as dom])
+  ```
+- Added `fulcro.client.alpha.localized-dom`. Works just like the new `dom`, but
+class keywords are localized to the *component* via fulcro-css rules.
 
 2.3.1
 -----